Split App Component out of index file
