+++ src/index.jsx
@@ -1,34 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 
-import Page from './components/Page';
-
-function App() {
-  const [state, setState] = useState({
-    count: 0,
-  });
-
-  const { count } = state;
-
-  function handleClick() {
-    setState({
-      count: count + 1,
-    });
-  }
-
-  function handleDigit(value) {
-    setState({
-      count: count + 1 * value,
-    });
-  }
-
-  return (
-    <Page
-      count={count}
-      onClick={handleClick}
-      handleDigit={handleDigit}
-    />
-  );
-}
+import App from './components/App';
 
 ReactDOM.render(<App />, document.getElementById('app'));
